+++ src/application/tests/testResponseBodyContent.test.ts
@@ -107,4 +107,28 @@ describe('testResponseBodyContent', () => {
     const pmTest = pmOperation.getTests()
     expect(pmTest.script.exec).toMatchSnapshot()
   })
+
+  it('should add content test for property check & result has minimum length value', async () => {
+    const contentTests = [
+      {
+        key: 'data[0].description',
+        minLength: 9
+      }
+    ]
+    pmOperation = testResponseBodyContent(contentTests, pmOperation)
+    const pmTest = pmOperation.getTests()
+    expect(pmTest.script.exec).toMatchSnapshot()
+  })
+
+  it('should add content test for property check & result has minimum length value', async () => {
+    const contentTests = [
+      {
+        key: 'data[0].description',
+        maxLength: 9
+      }
+    ]
+    pmOperation = testResponseBodyContent(contentTests, pmOperation)
+    const pmTest = pmOperation.getTests()
+    expect(pmTest.script.exec).toMatchSnapshot()
+  })
 })